This is a temporary fix to address issue #15 where batch process for a KPI yielding series values is not plotted.
I disabled the check for plottable KPIs in BatchProcessViewPage.js file.

The original design assumed that all KPIs are point values while the
intermediate results for calculting the KPIs can be plotted if they are
series data. The 'show_chart' attribute is set to true for aggregate KPIs with
intermediate results when api request is sent to the manager server.
However, when the KPIs are series data, such attribure does not exit.
This requires redesign of the view page and batch process api.